remove guide context menu
